Make devcontainer more stable and install required features
Use fixed devcontainer version.
This ensures the environment  is stable and always produces the same result.

Remove docker-outside-of-docker feature.
It does not work with podman and the evcc container image should be buildable with plain docker / podman.

Install Vite+ in devcontainer.
It is a new development tool for evcc and should be available in the devcontainer.